reworked sidebar, deleted ship & fleet , prep for fleet page
