--- src/Components/fleet/fleet.css
@@ -1,31 +1,0 @@
-.fleet {
-    height: 100%;
-    overflow-y: scroll;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem;
-    scrollbar-width: thin;
-    scrollbar-color: var(--bg-2) var(--bg-0);
-
-}
-
-.full_fleet {
-    padding: 0.5rem 1rem;
-    border-radius: 0.5rem;
-    background-color: var(--bg-2);
-    text-align: center;
-    text-decoration: none;
-    color: var(--accent-2);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0;
-    transition: 300ms;
-}
-
-.full_fleet:hover {
-    background-color: var(--bg-2-hover);
-    gap: 0.25rem;
-    color: var(--accent);
-}
