+++ src/Components/sidebar/sidebar.css
@@ -36,7 +36,128 @@
     font-weight: 400;
     color: var(--txt-muted);
     letter-spacing: 0.25rem;
+    white-space: nowrap;
 }
+
+.fleet {
+    height: 100%;
+    overflow-y: scroll;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
+    scrollbar-width: thin;
+    scrollbar-color: var(--bg-2) var(--bg-0);
+
+}
+
+.full_fleet:hover {
+    background-color: var(--bg-2-hover);
+    gap: 0.25rem;
+    color: var(--accent);
+}
+
+.full_fleet {
+    padding: 0.5rem 1rem;
+    border-radius: 0.5rem;
+    background-color: var(--bg-2);
+    text-align: center;
+    text-decoration: none;
+    color: var(--accent-2);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0;
+    transition: 300ms;
+}
+
+
+.fleetItem {
+    background-color: var(--bg-2);
+    border-radius: 0.5rem;
+    display: flex;
+    padding: 1rem;
+    flex-direction: column;
+    justify-content: center;
+    align-items: flex-start;
+    align-self: stretch;
+    gap: 0.5rem;
+    color: var(--txt);
+    text-decoration: none;
+    transition: 300ms;
+}
+
+.fleetItem:hover {
+    background-color: var(--bg-2-hover);
+}
+
+.fleetItem__header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    align-self: stretch;
+    gap: 1rem;
+}
+
+.fleetItem__symbol {
+    font-weight: 600;
+    color: var(--accent);
+    white-space: nowrap;
+    font-size: 1.25rem;
+}
+
+
+/* Useless now */
+.fleetItem__frameName {
+    font-size: 0.75rem;
+    font-weight: 500;
+    letter-spacing: 0.05rem;
+    background-color: var(--bg-0);
+    padding: 0.5rem;
+    border-radius: 0.25rem;
+}
+
+.fleetItem__status {}
+
+.fleetItem__location {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 0.25rem;
+}
+
+.fleetItem__locationIcon {}
+
+.fleetItem__waypointSymbol {}
+
+.fleetItem__travelStatus {
+    align-self: stretch;
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.fleetItem__timer {}
+
+.fleetItem__progressBar {
+    align-self: stretch;
+    height: 0.25rem;
+    border-radius: 100px;
+    background-color: var(--bg);
+    border: none;
+    outline: none;
+}
+
+.fleetItem__transitTiming {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+    align-self: stretch;
+}
+
+.fleetItem__departureTime {}
+
+.fleetItem__arrivalTime {}
 
 .logout {
     padding: 0.5rem 1rem;
