+++ src/Components/sidebar/sidebar.tsx
@@ -1,14 +1,86 @@
 import "./sidebar.css";
 import { Agent } from '../../Models/AgentInterface';
-import { Ship } from '../../Models/ShipInterface';
-import Fleet from '../fleet/fleet';
+import { Ship, Ship as ShipData } from '../../Models/ShipInterface';
 import User from '../user/user';
 import logo from "/assets/logo.svg";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../Context/auth/AuthContext";
+import { Link } from "react-router-dom";
+import ProgressBar from "../progressBar/progressBar";
 
-const Sidebar = ({ user, fleet }: { user: Agent | null, fleet: Ship[] | null }) => {
+import chevron from "/assets/icons/chevron_right.svg";
+import dockedIcon from "/assets/icons/docked.svg";
+import in_orbitIcon from "/assets/icons/in_orbit.svg";
+import in_transitIcon from "/assets/icons/in_transit.svg";
+import locationIcon from "/assets/icons/location.svg";
+
+const Sidebar = ({ user, fleet }: { user: Agent | null, fleet: ShipData[] | null }) => {
     const authContext = useContext(AuthContext);
+    const [timeUntilArrival, setTimeUntilArrival] = useState<number>(0);
+    const [flightProgress, setFlightProgress] = useState<number>(0);
+    let timer: NodeJS.Timeout;
+
+    const calculateTimeAndProgress = (ship: Ship) => {
+        let departureDate = new Date(ship.nav.route.departureTime);
+        let arrivalDate = new Date(ship.nav.route.arrival);
+
+        const currentTime = Date.now();
+        const timeUntilArrival = Math.floor((arrivalDate.getTime() - currentTime) / 1000);
+        const flightProgress = Math.floor(100 * (currentTime - departureDate.getTime()) / (arrivalDate.getTime() - departureDate.getTime()));
+
+        setTimeUntilArrival(timeUntilArrival);
+        setFlightProgress(flightProgress);
+
+        if (timeUntilArrival <= 0) {
+            clearInterval(timer);
+            let editedShipData = ship;
+            editedShipData.nav.status = "IN_ORBIT";
+            // setShipData(editedShipData);
+        }
+    };
+
+    const renderTransitInfos = (ship: ShipData) => {
+        let departureDate = new Date(ship.nav.route.departureTime);
+        let arrivalDate = new Date(ship.nav.route.arrival);
+
+        if (ship.nav.status === "IN_TRANSIT") {
+            return (
+                <div className='fleetItem__travelStatus'>
+                    <span className='fleetItem__timer'>{timeUntilArrival}s</span>
+                    <ProgressBar max={100} value={flightProgress} color='var(--confirm)'></ProgressBar>
+                    <div className='fleetItem__transitTiming'>
+                        <span className='fleetItem__departureTime'>{departureDate.toLocaleString("fr-FR").replace("/2023", "")}</span>
+                        <span className='fleetItem__arrivalTime'>{arrivalDate.toLocaleString("fr-FR").replace("/2023", "")}</span>
+                    </div>
+                </div>
+            )
+        }
+    }
+
+    useEffect(() => {
+        fleet?.map((ship: Ship) => {
+            calculateTimeAndProgress(ship);
+
+            timer = setInterval(() => {
+                calculateTimeAndProgress(ship);
+            }, 1000);
+        })
+
+        return () => clearInterval(timer); // Cleanup interval on component unmount
+    }, []); // Empty dependency array to run useEffect only once on mount
+
+    const renderStatusIcon = (ship: ShipData) => {
+        switch (ship.nav.status) {
+            case "DOCKED":
+                return <img className='fleetItem__status' src={dockedIcon} alt="DOCKED" />
+            case "IN_ORBIT":
+                return <img className='fleetItem__status' src={in_orbitIcon} alt="IN_ORBIT" />
+            case "IN_TRANSIT":
+                return <img className='fleetItem__status' src={in_transitIcon} alt="IN_TRANSIT" />
+            default:
+                return <img className='fleetItem__status' src="" alt="UNKNOWN" />
+        }
+    }
 
     return (
         <div className='sideBar'>
@@ -19,7 +91,31 @@ const Sidebar = ({ user, fleet }: { user: Agent | null, fleet: Ship[] | null }) 
                     <span className="hero__title--sub">Space Journey</span>
                 </div>
             </div>
-            <Fleet fleet={fleet} />
+            <Link to={"/fleet"} className="full_fleet">
+                <span>Full fleet</span>
+                <img src={chevron} />
+            </Link>
+            {fleet ? (
+                <div className="fleet">
+                    {fleet.map((ship: ShipData) => (
+                        <Link key={ship.symbol} to={"/fleet/" + ship.symbol} className='fleetItem'>
+                            <div className='fleetItem__header'>
+                                <span className='fleetItem__symbol'>{ship.registration.role.charAt(0).toUpperCase() + ship.registration.role.slice(1).toLowerCase()} {ship.frame.name}</span>
+                                {renderStatusIcon(ship)}
+                            </div>
+                            <div className='fleetItem__location'>
+                                <img className='fleetItem__locationIcon' src={locationIcon} alt="" />
+                                <span className='fleetItem__waypointSymbol'>{ship.nav.waypointSymbol}</span>
+                            </div>
+                            {renderTransitInfos(ship)}
+                        </Link>
+                    ))}
+                </div>
+            ) : (
+                <div className="fleet loading">
+                    <div className="loader"></div>
+                </div>
+            )}
             <User user={user} />
             <button onClick={() => { authContext.logout() }} className="logout">Log out</button>
         </div>
